refactor: extract debug controller
